connect baca nest api. remove example api implementation
