+++ src/hooks/forms/useSignInForm.ts
@@ -1,27 +1,25 @@
+import { useAuthControllerLogin } from '@baca/api/query/auth/auth'
 import { setToken } from '@baca/services'
 import { isSignedInAtom } from '@baca/store/auth'
 import { SignInFormValues } from '@baca/types/authForms'
-import { hapticImpact, wait } from '@baca/utils'
-import { isError } from '@tanstack/react-query'
+import { hapticImpact } from '@baca/utils'
 import { useSetAtom } from 'jotai'
-import { useState } from 'react'
 import { useForm } from 'react-hook-form'
-import { useTranslation } from 'react-i18next'
+import { notify } from 'react-native-notificated'
 
 const defaultValues: SignInFormValues = {
   // TODO: Reset this values when building production app
-  email: 'test@example.com',
-  password: '123456',
+  email: 'l.jeziorski+user@binarapps.com',
+  password: 'Qwerty1!',
   confirm: false,
 }
 
 export const useSignInForm = () => {
-  const [error, setError] = useState('')
-  const [isSubmitting, setIsSubmitting] = useState(false)
-
-  const { t } = useTranslation()
-
   const setIsSignedIn = useSetAtom(isSignedInAtom)
+
+  const { mutate: loginMutate, isLoading: isSubmitting } = useAuthControllerLogin<{
+    message: string
+  }>()
 
   const {
     control,
@@ -34,40 +32,30 @@ export const useSignInForm = () => {
   })
 
   const onSubmit = async (data: SignInFormValues) => {
-    try {
-      setIsSubmitting(true)
-      setError('')
-      // Errors are handled on UI side
-      // if you want to stop this function with error just throw new Error.
-      // Remember to pass readable error message for user, because this error will be displayed for him
-
-      // TODO: Add some backend call here, you can use react query for this
-      await wait(500)
-
-      if (data.email !== 'test@example.com' || data.password !== '123456') {
-        throw new Error('Incorrect email or password')
+    // Errors are handled on UI side
+    // if you want to stop this function with error just throw new Error.
+    // Remember to pass readable error message for user, because this error will be displayed for him
+    loginMutate(
+      { data },
+      // FIXME: add proper notification handling, generate some global config
+      {
+        onError: (e) => {
+          notify('error', { params: { title: 'ERROR', description: e?.message } })
+          hapticImpact()
+        },
+        onSuccess: async (response) => {
+          await setToken(response.accessToken)
+          setIsSignedIn(true)
+        },
       }
-      await setToken('token_here')
-      setIsSignedIn(true)
-    } catch (e) {
-      if (isError(e)) {
-        setError(e.message)
-      } else {
-        setError(t('errors.something_went_wrong'))
-      }
-      hapticImpact()
-    } finally {
-      setIsSubmitting(false)
-    }
+    )
   }
 
   return {
     submit: handleSubmit(onSubmit),
     isSubmitting,
-    setIsSubmitting,
-    setFocus,
     control,
     errors,
-    error,
+    setFocus,
   }
 }
