--- src/query-hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useGetData_EXAMPLE'
