+++ src/screens/DataFromBeScreen_EXAMPLE.tsx
@@ -1,8 +1,8 @@
+import { useArticlesControllerFindAll } from '@baca/api/query/articles/articles'
+import { ArticleEntity } from '@baca/api/types'
 import { Loader, Center, Text, Box, Spacer } from '@baca/components'
 import { useScreenOptions, useTranslation } from '@baca/hooks'
-import { useGetCity_EXAMPLE } from '@baca/query-hooks'
-import { TodoItem } from '@baca/types/todos'
-import React, { useCallback } from 'react'
+import { useCallback } from 'react'
 import { ListRenderItem, FlatList } from 'react-native'
 
 export const DataFromBeScreen_EXAMPLE = () => {
@@ -11,13 +11,15 @@ export const DataFromBeScreen_EXAMPLE = () => {
   useScreenOptions({
     title: t('navigation.screen_titles.data_from_be_screen_example'),
   })
-  const { dataList, isFetchedDataAfterMount } = useGetCity_EXAMPLE()
 
-  const renderItem: ListRenderItem<TodoItem> = useCallback(({ item: { title, id } }) => {
+  const { data: articles, isInitialLoading: isInitialLoadingArticles } =
+    useArticlesControllerFindAll({ page: 1, pageSize: 10 })
+
+  const renderItem: ListRenderItem<ArticleEntity> = useCallback(({ item: { id, title } }) => {
     return (
       <Box mb="1" bg="fg.brand.primary" borderRadius={2} m={2}>
-        <Text>{'title:' + title}</Text>
         <Text>{'id: ' + id}</Text>
+        <Text.Body mb={2}>{'title: ' + title}</Text.Body>
       </Box>
     )
   }, [])
@@ -30,7 +32,7 @@ export const DataFromBeScreen_EXAMPLE = () => {
         <Spacer y="1" />
         <FlatList
           ListEmptyComponent={
-            !isFetchedDataAfterMount ? (
+            !isInitialLoadingArticles ? (
               <Center height={400} flex={1}>
                 <Loader type="circle" />
               </Center>
@@ -38,7 +40,7 @@ export const DataFromBeScreen_EXAMPLE = () => {
               <Text>No data found</Text>
             )
           }
-          data={dataList}
+          data={articles}
           renderItem={renderItem}
         />
       </Center>
