+++ src/services/index.ts
@@ -1,5 +1,3 @@
-export * from './api'
 export * from './MockedServerService'
-export * from './requests'
 export * from './NotificationService'
 export * from './TokenService'
