--- src/services/requests/index.ts
@@ -1,1 +1,0 @@
-export * from './AnyService_EXAMPLE'
