--- src/types/todos.ts
@@ -1,8 +1,0 @@
-export type TodoItem = {
-  userId: number
-  id: number
-  title: string
-  completed: boolean
-}
-
-export type TodoList = TodoItem[]
